requirements.txt
annotated-types
anyio
apify_fingerprint_datapoints
attrs
Automat
beautifulsoup4
browserforge
bs4
cattrs
certifi
cffi
charset-normalizer
click
constantly
cryptography
cssselect
curl_cffi
defusedxml
filelock
greenlet
h11
httpcore
httpx
httpx-sse
hyperlink
idna
Incremental
iniconfig
itemadapter
itemloaders
jmespath
jsonschema
jsonschema-specifications
lxml
mcp
msgspec
orjson
packaging
parsel
patchright
pip
platformdirs
playwright
pluggy
Protego
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydantic-settings
PyDispatcher
pyee
Pygments
PyJWT
pyOpenSSL
pytest
pytest-base-url
pytest-playwright
python-dotenv
python-multipart
python-slugify
PyYAML
queuelib
referencing
requests
requests-cache
requests-file
rpds-py
scrapling
service-identity
soupsieve
sse-starlette
starlette
text-unidecode
tld
tldextract
Twisted
typing_extensions
typing-inspection
url-normalize
urllib3
uvicorn
w3lib
zope.interface
